refactor: 입력, 출력 메서드 Controller로 이동 입력하는 행위와 출력하는 행위 모두 사용자와 직접적으로 데이터나 값을 주고받아야 하기에, Service에서 Controller로 이동
