+++ src/main/java/racingcar/RacingController.java
@@ -1,5 +1,7 @@
 package racingcar;
 
+import camp.nextstep.edu.missionutils.Console;
+
 import java.util.*;
 
 public class RacingController {
@@ -18,7 +20,13 @@ public class RacingController {
     }
 
     public List<Car> saveCarNames() {
-        return racingService.parseCarNamesToArray(racingService.inputCarNames());
+        return racingService.parseCarNamesToArray(inputCarNames());
+    }
+
+    public String inputCarNames() {
+        System.out.println("경주할 자동차 이름을 입력하세요.(이름은 쉼표(,) 기준으로 구분)");
+
+        return Console.readLine();
     }
 
     public void printResultText() {
@@ -26,14 +34,35 @@ public class RacingController {
     }
 
     public void judgeCarMove() {
-        Integer tryNumber = racingService.inputNumberOfTries();
+        Integer tryNumber = inputNumberOfTries();
 
         for(int i=0; i<tryNumber; i++) {
             if(racingService.decideAdvance()) {
                 racingService.increaseCarCount();
             }
-            racingService.printCountView();
+            printCountView();
         }
+    }
+
+    public Integer inputNumberOfTries() {
+        System.out.println("시도할 회수는 몇회인가요?");
+
+        return Integer.parseInt(Console.readLine());
+    }
+
+    public void printCountView() {
+        for (Car car : racingService.getCarList()) {
+            String carName = car.getCarName();
+            int count = car.getAdvanceCount();
+
+            StringBuilder sb = new StringBuilder();
+            for (int i = 0; i < count; i++) {
+                sb.append("-");
+            }
+
+            System.out.println(carName + " : " + sb.toString());
+        }
+        System.out.println();
     }
 
     public void printWinnerResult() {
